Refactor: move the to-usd calculations to a new module tousd.py
